now it outputs all 8 vertices.
at this point, it does so by using a somewhat obscure method. so for
each cell that has intersections on its minimal axis, it will also mark
its negative-adjacent cells to generate their respective minimizers.
likely we still have to switch to the classic 3d array method for
storing the hermite data and do the full scan to generate
minimizers(vertices).